requirements.txt
albumentations==1.4.1
diffusers==0.9.0
kornia==0.7.1
natsort==8.4.0
numpy==1.26.4
opencv_python==4.9.0.80
opencv_python_headless==4.9.0.80
Pillow==9.3.0
Pillow==10.2.0
pretrainedmodels==0.7.4
protobuf==4.25.3
torch==2.1.2+cu118
torchvision==0.16.2+cu118
tqdm==4.66.1
